vscode: increasing file explorer/workspace indentation to 21 from 16.
some file icons are not left aligned with the dir above, this make it
closer to be aligned on the current theme (monokai pro), such as python,
toml, etc. but then other icons (which are thinner) are more indented
to the left and are out of left alignment with the dir above.

sticking with the left alignment of the bigger icons for now.